The providers approach is lengthy and tiresome. Removed relevant classes.
Instead, 'dependencies' should be resolvable, as they were before, however a form of dependency must be resolvable using a provider. In this way, providers act as a convenient interface for building new objects, to be used in application or library code, rather than as the backbone for dependency injection on the whole.