Implemented prototype and singleton scopes, closes #3. Also added typedef for providers, working on #8.
Scopes individually aid in creating objects. They make use of providers to decouple from the object graph factory.
The prototype scope returns the result of its provider.
The singleton scope first returns the result, and then returns the first result for all subsequent calls, creating a one of a kind.